YT_Chatbot/requirements.txt
streamlit
youtube-transcript-api==0.6.2
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
huggingface-hub
